+++ src/core/utils.jl
@@ -163,7 +163,8 @@ function dictkeys_tosymbols(d::Dict)
             "monthly_totals_kwh",
             "prod_factor_series_kw", 
             "monthly_energy_rates", "monthly_demand_rates",
-            "wholesale_rate", "blended_doe_reference_percents"
+            "wholesale_rate", "blended_doe_reference_percents",
+            "coincident_peak_load_charge_per_kw"
             ] && !isnothing(v)
             try
                 v = convert(Array{Real, 1}, v)
@@ -178,6 +179,15 @@ function dictkeys_tosymbols(d::Dict)
                 v = convert(Array{String, 1}, v)
             catch
                 @warn "Unable to convert $k to an Array{String, 1}"
+            end
+        end
+        if k in [
+            "coincident_peak_load_active_timesteps"
+        ]
+            try
+                v = convert(Vector{Vector{Int64}}, v)
+            catch
+                @warn "Unable to convert $k to a Vector{Vector{Int64}}"
             end
         end
         d2[Symbol(k)] = v